+++ src/main/kotlin/com/whatever/config/openfeign/FeignClientGlobalConfig.kt
@@ -1,0 +1,15 @@
+package com.whatever.config.openfeign
+
+import feign.Logger
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+
+@Configuration
+class FeignClientGlobalConfig {
+
+    @Bean
+    fun feignLoggerLevel(): Logger.Level {
+        return Logger.Level.FULL
+    }
+
+}
