+++ src/main/kotlin/com/whatever/config/openfeign/KakaoKauthConfig.kt
@@ -1,0 +1,14 @@
+package com.whatever.config.openfeign
+
+import feign.codec.Encoder
+import feign.form.FormEncoder
+import org.springframework.context.annotation.Bean
+
+class KakaoKauthConfig {
+
+    @Bean
+    fun formEncoder(): Encoder {
+        return FormEncoder()
+    }
+
+}
